lzw: move reader to asm

+++ lzw/decode_table.h
@@ -13,7 +13,7 @@ struct decode_table
 };
 
 void decode_table_init(struct decode_table *table);
-bool decode_table_contains(struct decode_table const *table, int16_t code);
+bool decode_table_contains(struct decode_table const *table, uint16_t code);
 int16_t decode_table_append(struct decode_table *table, int16_t code, uint8_t byte);
-int16_t decode_table_get_first_byte(struct decode_table const *table, int16_t code);
-void decode_table_write_bytes(uint8_t **w, int16_t code, struct decode_table const *table);
+uint8_t decode_table_get_first_byte(struct decode_table const *table, uint16_t code);
+void decode_table_write_bytes(uint8_t **w, uint16_t code, struct decode_table const *table);
